+++ native/storybook/storybook.js
@@ -51,6 +51,7 @@ addDecorator(CenterDecorator);
 /* eslint-disable global-require */
 configure(() => {
   require('../packages/react-native-bpk-component-animate-height/stories');
+  require('../packages/react-native-bpk-component-banner-alert/stories');
   require('../packages/react-native-bpk-component-button/stories');
   require('../packages/react-native-bpk-component-card/stories');
   require('../packages/react-native-bpk-component-icon/stories');